feat: implemented ringing screen and alarm scheduling
Added a RingingScreen which is triggered by an AlarmReceiver upon alarm time.
Implemented deep linking for opening the RingingScreen via a notification.
Created a NotificationHelper to show a fullscreen notification when the alarm is triggered.
Scheduled a sample alarm using NotificationHelper to test the ringing functionality.
Created ScheduleAlarmUseCase and CancelAlarmUseCase to with empty logic